requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
db-dtypes==1.3.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.1
openai==1.58.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
